refactor: Move service interfaces to 'impl' package
- Moved service interfaces to the 'impl' package to better organize the
codebase
- The interface methods remain unchanged, only their location has been
updated
