+++ backend/src/main/java/com/chatop/rental_portal_backend/services/impl/IJwtService.java
@@ -1,4 +1,4 @@
-package com.chatop.rental_portal_backend.services;
+package com.chatop.rental_portal_backend.services.impl;
 
 import org.springframework.security.core.Authentication;
 import org.springframework.security.oauth2.jwt.Jwt;
@@ -6,8 +6,5 @@ import org.springframework.security.oauth2.jwt.Jwt;
 public interface IJwtService {
     String generateToken(Authentication authentication);
 
-    boolean validateToken(String token);
-
     Jwt decodeToken(String token);
-
 }
