+++ backend/src/main/java/com/chatop/rental_portal_backend/services/impl/IMessageService.java
@@ -1,4 +1,4 @@
-package com.chatop.rental_portal_backend.services;
+package com.chatop.rental_portal_backend.services.impl;
 
 import com.chatop.rental_portal_backend.dto.MessageRequestDTO;
